requirements.txt
alembic==1.4.2
amqp==2.5.2
anyjson==0.3.3
astroid==2.4.1
autopep8==1.5.2
billiard==3.6.3.0
bleach==3.1.5
blinker==1.4
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.1
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-PageDown==0.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-SSLify==0.1.5
Flask-WTF==0.14.3
ForgeryPy==0.1
html5lib==1.0.1
httpie==2.1.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.8
lazy-object-proxy==1.4.3
Mako==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.3
pycodestyle==2.6.0
Pygments==2.6.1
pylint==2.5.2
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
redis==3.5.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
SQLAlchemy==1.3.17
toml==0.10.1
urllib3==1.25.9
vine==1.3.0
visitor==0.1.3
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
